+++ tests/integration/test_sub_clients/test_sync_clients/test_charge.py
@@ -18,7 +18,7 @@ class ChargeClientTestCase(TestCase):
         bank_data = {"code": "057", "account_number": "0000000000"}
         response = self.client.charge(
             email="coyotedevmail@gmail.com",
-            amount=1000,
+            amount=10000,
             bank=bank_data,
         )
         self.assertEqual(response.status_code, httpx.codes.OK)
@@ -45,10 +45,12 @@ class ChargeClientTestCase(TestCase):
         )
         self.assertEqual(response.message, "Charge attempted")
 
+    @skip("incomplete test")
     def test_can_submit_birthday(self) -> None:
         response = self.client.submit_birthday(
-            birthday="1999-04-29", reference="kv7ecgjrit1fxgs"
+            birthday="1999-04-29", reference="2wy4icde635aofc"
         )
+        print(response)
         self.assertTrue(response.status)
         self.assertEqual(response.message, "Charge attempted")
 
@@ -64,6 +66,7 @@ class ChargeClientTestCase(TestCase):
         )
         self.assertEqual(response.message, "Charge attempted")
 
+    @skip("incomplete test")
     def test_can_check_pending_charge(self) -> None:
         response = self.client.check_pending_charge(reference="kv7ecgjrit1fxgs")
         self.assertTrue(response.status)